Combine LegacyComboCounter and ComboCounter classes
